Now with hot POST action
Can now add containers and key/values with tags.

Fixed an issue in the Container model, there's no TAG there, but was
incorrectly defined in the model.

In the gets, sorted the URL/Query Params in order to minimize the amount
of items cached, and to ensure a consistent fashion to enable cache
expiration.

Added the supported post actions, for adding a container to the tree,
and for adding key value pairs into a container, it's primarily in
posts.py. This is where all of the cache invalidation logic is coded.
Might need to move this out to a common/util location, since PUTS will
also need to invalidate the cache the same way.

routes.py, removed extraneous debug logging