Documentation, Cleanup, Formatting & Refactored GetValues
Full pep8 compliance (even if it makes me ill... I mean I <3 78 chars
per line)

Found while doing the documentation that get_values was not properly
checking for duplicate items. Fixed this issue, and completely rewrote
the way that particular call is handled. It no longer returns a list of
value objects which are encrypted, but returns a list of dictionaries
containing all the value information, decrypted.

Cleaned up comments and added docstring's for all methods.