refactored, more or less rewritten
Wanted to keep it simpler, and the way I previously had the crud config
objects didn't make much sense. This is rewritten, and tested for higher
performance. Much of the system remains the same, and the actual code
differences only took a few hours. Overall the system should be much
easier to expand, and scale.

One of the notable differenes is encryption of database values is no
longer implemented, this was a design decision and may be re-implemented
in the future